refactor: consolidate MySqlDeployerValidationTest into MySqlDeployerTest
Merged all tests from MySqlDeployerValidationTest (1030 lines) into
MySqlDeployerTest, removing true duplicates (same-assertion null-database
and restore no-op tests) and consolidating shared helper methods
(stubConnection, stubConnectionWithStatement, stubDefaultRowType).
